backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-multipart
dotenv
